requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pillow==10.4.0
pypdf>=4.0.0
opencv-python-headless>=4.8.0
pytesseract>=0.3.10
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
imagehash>=4.3.1
rapidfuzz>=3.5.0
gspread>=5.12.0
google-auth>=2.23.0
pytest==8.3.2
httpx==0.27.2
